refactor: centralize CostItem type
